+++ src/main/clojure/com/github/clojure_repl/intellij/action/eval.clj
@@ -5,6 +5,7 @@
    [com.github.clojure-repl.intellij.nrepl :as nrepl]
    [com.github.clojure-repl.intellij.parser :as parser]
    [com.github.clojure-repl.intellij.ui.hint :as ui.hint]
+   [com.github.clojure-repl.intellij.ui.inlay-hint :as ui.inlay-hint]
    [com.github.clojure-repl.intellij.ui.repl :as ui.repl]
    [com.github.ericdallo.clj4intellij.app-manager :as app-manager]
    [com.github.ericdallo.clj4intellij.tasks :as tasks]
@@ -19,7 +20,7 @@
 
 (set! *warn-on-reflection* true)
 
-(defn ^:private eval-action [^AnActionEvent event loading-msg eval-fn success-msg-fn]
+(defn ^:private eval-action [^AnActionEvent event loading-msg eval-fn success-msg-fn {:keys [inlay-hint-feedback?]}]
   (when-let [editor ^Editor (.getData event CommonDataKeys/EDITOR_EVEN_IF_INACTIVE)]
 
     (let [project (.getProject editor)]
@@ -40,7 +41,9 @@
                    (ui.hint/show-error {:message (str "Namespace not found: " (:ns response)) :editor editor})
 
                    :else
-                   (ui.hint/show-repl-info :message (success-msg-fn response) :editor editor)))}))))
+                   (if inlay-hint-feedback?
+                     (ui.inlay-hint/show-code (success-msg-fn response) editor)
+                     (ui.hint/show-repl-info :message (success-msg-fn response) :editor editor))))}))))
         (ui.hint/show-error :message "No REPL connected" :editor editor)))))
 
 (defn load-file-action [^AnActionEvent event]
@@ -51,7 +54,8 @@
      (fn [^Editor editor]
        (nrepl/load-file (.getProject editor) editor virtual-file))
      (fn [_response]
-       (str "Loaded file " (.getPath ^VirtualFile virtual-file))))))
+       (str "Loaded file " (.getPath ^VirtualFile virtual-file)))
+     {})))
 
 (defn eval-last-sexpr-action [^AnActionEvent event]
   (when-let [editor ^Editor (.getData event CommonDataKeys/EDITOR_EVEN_IF_INACTIVE)]
@@ -67,7 +71,8 @@
                ns (some-> (parser/find-namespace root-zloc) z/string parser/remove-metadata)]
            (nrepl/eval {:project (.getProject editor) :code code :ns ns})))
        (fn [response]
-         (string/join "\n" (:value response)))))))
+         (string/join "\n" (:value response)))
+       {:inlay-hint-feedback? true}))))
 
 (defn eval-defun-action [^AnActionEvent event]
   (eval-action
@@ -83,7 +88,8 @@
            ns (some-> (parser/find-namespace root-zloc) z/string parser/remove-metadata)]
        (nrepl/eval {:project (.getProject editor) :code code :ns ns})))
    (fn [response]
-     (string/join "\n" (:value response)))))
+     (string/join "\n" (:value response)))
+   {:inlay-hint-feedback? true}))
 
 (defn clear-repl-output-action [^AnActionEvent event]
   (let [editor ^Editor (.getData event CommonDataKeys/EDITOR_EVEN_IF_INACTIVE)
@@ -114,7 +120,8 @@
            namespace (parser/remove-metadata (z/string zloc))]
        (nrepl/eval {:project (.getProject editor) :code (format "(in-ns '%s)" namespace) :ns namespace})))
    (fn [response]
-     (string/join "\n" (:value response)))))
+     (string/join "\n" (:value response)))
+   {}))
 
 (defn refresh-all-action [^AnActionEvent event]
   (eval-action
@@ -125,7 +132,8 @@
    (fn [response]
      (if (contains? (:status response) "ok")
        "Refreshed sucessfully"
-       (str "Refresh failed:\n" (:error response))))))
+       (str "Refresh failed:\n" (:error response))))
+   {}))
 
 (defn refresh-changed-action [^AnActionEvent event]
   (eval-action
@@ -136,4 +144,5 @@
    (fn [response]
      (if (contains? (:status response) "ok")
        "Refreshed sucessfully"
-       (str "Refresh failed:\n" (:error response))))))
+       (str "Refresh failed:\n" (:error response))))
+   {}))