+++ CHANGELOG.md
@@ -5,6 +5,7 @@
  - Add icons of REPL commands to REPL window (clear and entry history navigation)
  - Drop support of older IntelliJ versions (2021/2022). Now requires minimum IntelliJ 2023.3 (Build 233)
  - Fix namespace-not-found error handling while evaluating
+ - Add eval inlay hint support. #106
 
 ## 1.6.2