+++ CHANGELOG.md
@@ -2,11 +2,12 @@
 
 ## [Unreleased]
 
- - Add icons of REPL commands to REPL window (clear and entry history navigation)
+ - Add icons of REPL commands to REPL window (clear and entry history navigation). #99
  - Drop support of older IntelliJ versions (2021/2022). Now requires minimum IntelliJ 2023.3 (Build 233)
- - Fix namespace-not-found error handling while evaluating
+ - Fix namespace-not-found error handling. Now shows a message to the user. #107
  - Add eval inlay hint support. #106
  - Add action to interrupt evals on the REPL session (`shift alt R` + `shift alt S`).  #104
+ - Add color settings page for customization of some tokens.
 
 ## 1.6.2